architect, bard
