+++ src/model/card.ts
@@ -44,6 +44,8 @@ import {
 } from "./gameText";
 import { assertUnreachable } from "../utils";
 
+import { architect, bard } from "./cards/base";
+
 type NumWorkersInnerFn = (cardOwner: Player) => number;
 type ProductionInnerFn = (
   gameState: GameState,
@@ -465,86 +467,8 @@ export class Card<TCardType extends CardType = CardType>
 }
 
 const CARD_REGISTRY: Record<CardName, Card> = {
-  [CardName.ARCHITECT]: new Card({
-    name: CardName.ARCHITECT,
-    numInDeck: 2,
-    cardType: CardType.PROSPERITY,
-    baseCost: { [ResourceType.BERRY]: 4 },
-    baseVP: 2,
-    isUnique: true,
-    isConstruction: false,
-    associatedCard: CardName.CRANE,
-    cardDescription: toGameText([
-      { type: "points", value: 1 },
-      " for each of your unused RESIN and PEBBLE, to a maximum of 6.",
-    ]),
-    // 1 point per rock and pebble, up to 6 pts
-    pointsInner: (gameState: GameState, playerId: string) => {
-      const player = gameState.getPlayer(playerId);
-      const numPebblesAndResin =
-        player.getNumResourcesByType(ResourceType.PEBBLE) +
-        player.getNumResourcesByType(ResourceType.RESIN);
-      return numPebblesAndResin > 6 ? 6 : numPebblesAndResin;
-    },
-  }),
-  [CardName.BARD]: new Card({
-    name: CardName.BARD,
-    numInDeck: 2,
-    cardType: CardType.TRAVELER,
-    baseCost: { [ResourceType.BERRY]: 3 },
-    baseVP: 0,
-    isUnique: true,
-    isConstruction: false,
-    associatedCard: CardName.THEATRE,
-    cardDescription: toGameText(
-      "You may discard up to 5 CARD to gain 1 VP each."
-    ),
-    playInner: (gameState: GameState, gameInput: GameInput) => {
-      const player = gameState.getActivePlayer();
-      if (gameInput.inputType === GameInputType.PLAY_CARD) {
-        gameState.pendingGameInputs.push({
-          inputType: GameInputType.DISCARD_CARDS,
-          label: "Discard up to 5 CARD to gain 1 VP each",
-          prevInputType: gameInput.inputType,
-          minCards: 0,
-          maxCards: 5,
-          cardContext: CardName.BARD,
-          clientOptions: {
-            cardsToDiscard: [],
-          },
-        });
-      } else if (
-        gameInput.inputType === GameInputType.DISCARD_CARDS &&
-        gameInput.cardContext === CardName.BARD
-      ) {
-        if (gameInput.clientOptions?.cardsToDiscard) {
-          if (gameInput.clientOptions.cardsToDiscard.length > 5) {
-            throw new Error("Discarding too many cards");
-          }
-          const numDiscarded = gameInput.clientOptions.cardsToDiscard.length;
-          if (numDiscarded === 0 && gameInput.isAutoAdvancedInput) {
-            gameState.addGameLogFromCard(CardName.BARD, [
-              player,
-              ` has no CARD to discard.`,
-            ]);
-          } else {
-            gameState.addGameLogFromCard(CardName.BARD, [
-              player,
-              ` discarded ${numDiscarded} CARD to gain ${numDiscarded} VP.`,
-            ]);
-          }
-          gameInput.clientOptions.cardsToDiscard.forEach((cardName) => {
-            player.removeCardFromHand(gameState, cardName);
-          });
-          player.gainResources(gameState, {
-            [ResourceType.VP]: numDiscarded,
-          });
-        }
-      } else {
-        throw new Error(`Unexpected input type ${gameInput.inputType}`);
-      }
-    },
-  }),
+  [CardName.ARCHITECT]: architect(),
+  [CardName.BARD]: bard(),
   [CardName.BARGE_TOAD]: new Card({
     name: CardName.BARGE_TOAD,
     numInDeck: 3,
