+++ src/model/cards/base/index.ts
@@ -1,0 +1,2 @@
+export { architect } from "./architect";
+export { bard } from "./bard";
